sepreate Project class from main script to run
